update README.md and add TimeStamp component
