+++ README.md
@@ -1,1 +1,26 @@
-# Happy Thoughts
+Happy Thoughts 
+
+This is a small React project where users can write “happy thoughts,” like posts, and see when each message was created.
+Users can write a message and post it.
+The message appears instantly in a card.
+
+❤️ Like a message
+
+Every card has a like button that switches between a filled and empty heart.
+The like count increases or decreases depending on the user's action.
+
+⏱ Timestamp
+
+Each message shows how long ago it was posted (e.g. “20 sec ago”).
+
+*  Technologies used
+
+React
+
+Vite
+
+JavaScript (ES6+)
+
+Tailwind utility classes (flex, spacing, alignment)
+
+useState / useEffect
